// src/_requests/Maja/models/NursesCreatePermissionsRequestBody.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { NursesCreatePermissionsRequestBodyPermission } from './NursesCreatePermissionsRequestBodyPermission';

export type NursesCreatePermissionsRequestBody = {
    expire_date?: string;
    nurseId?: number;
    permission?: NursesCreatePermissionsRequestBodyPermission;
};